requirements.txt
pyrtlsdr[lib]
scipy
setuptools
websockets
